pyproject.toml
[project]
name = "alascca-classify"
version = "0.1.0"
description = "Classification of PI3K pathway alterations in colorectal cancer for adjuvant aspirin eligibility per the ALASCCA trial"
readme = "README.md"
authors = [
    { name = "viktorlj", email = "viktor.ljungstrom@igp.uu.se" }
]
requires-python = ">=3.11"
dependencies = [
    "polars>=1.0",
    "pydantic>=2.0",
    "typer>=0.9",
    "rich>=13.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "fpdf2>=2.7",
]

[project.scripts]
alascca-classify = "alascca_classify.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/alascca_classify"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "marimo>=0.6",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
